% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylMutationAssociation.R
\name{methylMutationAssociation}
\alias{methylMutationAssociation}
\title{Methylation-Mutation Association}
\usage{
methylMutationAssociation(
  methylationCancer,
  methylationNormal,
  geneCancer,
  mutationData
)
}
\arguments{
\item{methylationCancer}{Matrix containing the methylation data
belonging to a sample representative of disease being studied}

\item{methylationNormal}{Matrix containing the methylation data
belonging to a sample representative of baseline methylation}

\item{geneCancer}{Matrix containing the gene expression data belonging to
a sample representative of the disease being studied}

\item{mutationData}{List organizing genes based on the category of their
mutation}
}
\description{
A function that generates a plot highlighting the relationship between
mutation category and differential DNA methylation.
}
\examples{
library(MethylMix)
data(METcancer)
data(METnormal)
data(GEcancer)
data(mutationExample)
methylMutationAssociation(METcancer, METnormal, GEcancer, mutationExample)


}
\references{
Gevaert O (2022). MethylMix: Identifying methylation driven
cancer genes. R package version 2.26.0.

R Core Team (2022). R: A language and environment for statistical computing.
R Foundation for Statistical Computing, Vienna, Austria.
URL https://www.R-project.org/
}
