% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylOverrep.R
\name{methylOverrep}
\alias{methylOverrep}
\title{Differential DNA Methylation Overrepresentation Analysis}
\usage{
methylOverrep(
  methylationCancer,
  methylationNormal,
  geneCancer,
  pValue = 0.05,
  qValue = 0.1
)
}
\arguments{
\item{methylationCancer}{Matrix containing the methylation data
belonging to a sample representative of disease being studied}

\item{methylationNormal}{Matrix containing the methylation data
belonging to a sample representative of baseline methylation}

\item{geneCancer}{Matrix containing the gene expression data belonging to
a sample representative of the disease being studied}

\item{pValue}{double representing the cutoff for how low a p-value must be
for a biological pathway to be considered enriched}

\item{qValue}{double representing the cutoff for how low a q-value must be
for a biological pathway to be considered enriched}
}
\value{
Returns a bar plot that summarizes the findings of the
overrepresentation analysis performed using the driver genes derived from
DNA methylation samples
}
\description{
A function that generates a plot summarizing the findings of an
overrepresentation analysis performed using the driver genes derived from
DNA methylation samples.
}
\examples{
library(MethylMix)
data(METcancer)
data(METnormal)
data(GEcancer)
resPlot <- methylOverrep(METcancer, METnormal, GEcancer, 1, 1)
resPlot

}
\references{
Carlson M (2022). org.Hs.eg.db: Genome wide annotation for Human. R package
version 3.15.0.

Gevaert O (2022). MethylMix: Identifying methylation driven cancer genes. R
package version 2.26.0.

R Core Team (2022). R: A language and environment for statistical computing.
R Foundation for Statistical Computing, Vienna, Austria.
URL https://www.R-project.org/.

T Wu, E Hu, S Xu, M Chen, P Guo, Z Dai, T Feng, L Zhou, W Tang, L Zhan,
X Fu, S Liu, X Bo, and G Yu. clusterProfiler 4.0: A universal enrichment
tool for interpreting omics data. The Innovation. 2021, 2(3):100141
}
