% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutationExample}
\alias{mutationExample}
\title{Gene-Mutation list}
\format{
An object of class \code{list} of length 4.
}
\usage{
mutationExample
}
\description{
An example of a list of genes filed under their respective mutation categories
}
\examples{
 \dontrun{
  mutationExample
 }
}
\keyword{datasets}
