% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMethylBioPaths.R
\name{runMethylBioPaths}
\alias{runMethylBioPaths}
\title{Launch Shiny App for MethylBioPaths}
\usage{
runMethylBioPaths()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the Shiny app for MethylBioPaths
The purpose of this app is only to illustrate how a Shiny app works. The
code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

MethylBioPaths::runMethylBioPaths()
}

}
\references{
Chang W, Cheng J, Allaire J, Sievert C, Schloerke B, Xie Y, Allen J,
McPherson J, Dipert A, Borges B (2022). _shiny: Web Application Framework
for R_. R package version 1.7.3, <https://CRAN.R-project.org/package=shiny>.
}
