% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutationExample}
\alias{mutationExample}
\title{Gene-Mutation list}
\format{
An object of class \code{list} of length 4.
}
\usage{
mutationExample
}
\description{
An example of a list of genes filed under their respective mutation
categories. The example is an R list object where the category each gene
belongs to is defined in the names of the List. Each entry in the List is a
vector of characters containing a series of gene names.
}
\examples{
 \dontrun{
  mutationExample
 }
}
\keyword{datasets}
